requirements.txt
django>=5.0
django-cors-headers
django-rest-framework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
requests
whitenoise
